Extracted visits charts elements into reusable component
